files now have owners, only owners or admins can delete

+++ apps/fileupload/views.py
@@ -1,5 +1,11 @@
 from itwishlist.apps.fileupload.models import File
+from itwishlist.apps.fileupload.forms import FileUploadForm
+
 from django.views.generic import CreateView, DeleteView, DetailView, ListView
+
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
+from django.utils.decorators import method_decorator
 
 from django.template.defaultfilters import slugify
 from django.contrib.auth.models import User
@@ -9,6 +15,7 @@ from django.contrib import messages
 from django.http import HttpResponse, HttpResponseRedirect
 from django.utils import simplejson
 from django.core.urlresolvers import reverse
+from django.shortcuts import get_object_or_404
 
 from django.conf import settings
 
@@ -18,12 +25,19 @@ def response_mimetype(request):
     else:
         return "text/plain"
 
+class LoginRequiredMixin(object):
+    """ View mixin which verifies that the user has authenticated. """
+
+    @method_decorator(login_required)
+    def dispatch(self, request, *args, **kwargs):
+        return super(LoginRequiredMixin, self).dispatch(request, *args, **kwargs)
+
 # class FileMixin(object):
 #     model = File
 #     def get_success_url(self):
 #         return reverse('upload-new')
-#     # def get_queryset(self):
-#     #     return File.objects.filter(owned_by=self.request.user)
+#     def get_queryset(self):
+#         return File.objects.filter(uploaded_by=self.request.user)
 # 
 # class FileDetailView(FileMixin, DetailView):
 #     pass
@@ -38,21 +52,43 @@ class FileListView(ListView):
     
     template_name = 'fileupload/file_list.html'
 
-class FileCreateView(CreateView):
+
+class FileCreateView(LoginRequiredMixin, CreateView):
+
     model = File
-    
-    def form_valid(self, form):
-        self.object = form.save()
-        f = self.request.FILES.get('file')
-        data = [{'name': f.name, 'url': settings.MEDIA_URL + "files/" + f.name.replace(" ", "_"), 'thumbnail_url': settings.MEDIA_URL + "files/" + f.name.replace(" ", "_"), 'delete_url': reverse('upload-delete', args=[self.object.id]), 'delete_type': "DELETE"}]
-        response = JSONResponse(data, {}, response_mimetype(self.request))
-        response['Content-Disposition'] = 'inline; filename=files.json'
-        return response
+    form_class = FileUploadForm        
 
     def get_context_data(self, **kwargs):
         context = super(FileCreateView, self).get_context_data(**kwargs)
         context['files'] = File.objects.all()
         return context
+    
+    def form_valid(self, form):
+        self.object = form.save(commit=False)
+        self.object.uploaded_by = self.request.user
+        self.object = form.save()
+        f = self.request.FILES.get('file')
+        data = [{'name': f.name, 
+                 # 'url': settings.MEDIA_URL + "files/" + f.name.replace(" ", "_"), 
+                 # 'uploaded_by': f.uploaded_by,
+                 'sitename': settings.SITE_NAME,
+                 'url': reverse('upload-detail', args=[self.object.id, self.object.slug]), 
+                 # 'thumbnail_url': settings.MEDIA_URL + "files/" + f.name.replace(" ", "_"), 
+                 'delete_url': reverse('upload-delete', args=[self.object.id]), 
+                 'delete_type': "DELETE"}]
+        response = JSONResponse(data, {}, response_mimetype(self.request))
+        response['Content-Disposition'] = 'inline; filename=files.json'
+        return response
+    
+    # def form_invalid(self, form):
+    #     print form
+    #     return super(FileCreateView, self).form_invalid(form)
+            
+    # def dispatch(self, request, *args, **kwargs):
+    #     self.file = get_object_or_404(File)
+    #     if self.file.uploaded_by == request.user:
+    #         return super(FileCreateView, self).dispatch(request, *args, **kwargs)
+    #     raise PermissionDenied
 
 
 class FileDeleteView(DeleteView):
@@ -63,7 +99,10 @@ class FileDeleteView(DeleteView):
         This does not actually delete the file, only the database record.  But
         that is easy to implement.
         """
+        # http://stackoverflow.com/a/5532445/412329
         self.object = self.get_object()
+        if not self.object.uploaded_by == self.request.user:
+            raise PermissionDenied
         self.object.delete()
         messages.add_message(request, messages.SUCCESS, message=("Successfully deleted file."))
         if request.is_ajax():
